+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # FUSE for Rust - Changelog
 
 ## 0.6.0 - UNRELEASED
-* make `spawn_mount()` safe
+* Make `spawn_mount()` safe
+* Add support for `fallocate()` and `ioctl()`
 
 ## 0.5.0 - 2020-10-17